Refactor the placement of the feature and headline folders to be now placed into the body folder
